Add the additional configuration selection page.
Added the ability to select to add the additional configuration of Debug
AVaRICE at the end of the project creation.

It is optional, so the FINISH button can be hit without seeing this
page.

The configs are now an arraylist of ConfigDescriptors (A simple struct-
like class for holding name and toolchain details) from the
BuildConfigurationsPage. The struct also contains a flag that indicates
to use debug compiler settings or not.