#424 Importing both a ArduinoIDE and ArduinoEclipseplugin project
If the file already exists it is not over written. 
This solution Does not work with the samples for arduino IDE imported
projects. You need to select the ino or cpp option to work
I also made sure you can import a arduino project. That is you can now
delete a project and not have the folders removed and directly reimport
the project.