#439 this was sort of a racing condition
What happened what that the boards.txt is processed line by line. In the
esp file the upload.tool is specified 3 times for the tested board.
Once as upload.tool and 2 times for the upload protocol.
Apparently the upload.tool was processed after the menu.
So one could say that this is a bug in the boards.txt file.
However I catered for this situation.