ATO is now working with security
Note that the plugin reads the pwd from the source code and stores it in
the security cache after the upload.
The pwd used is the pwd found in the stored cache unless no pwd was
found there.
This should support changing and removing pwd's without ever having to
type one

I also added "never and always" as options in the dialog box "build
before upload"