Added changes to the generator that had been done by the cdt team.
Note that I tried to switch to a more OO approach by overloading
GnuMakefileGenerator. However that didn't work as I need access to
private members (variables). As these variable members are initialized
by code I do not need to overload but I do need access to it I don not
know how to implement this.