After I fixed #393 I wanted to test it with the unit test program.
Out came a bug that if a recipe is left empty it causes problems.
So I fixed that.
It seemed that one of the boards changed its names in the menu options.
So I changed the option content from <menuName,MenuItemName> to
<menuID,MenuItemID>. 
Doing so I noticed the info was stored at 2 places (some key thingy and
in the environment variables). I went for only environment variables.
However environment variables are case insensetive -at least in the
implementation of Sloeber-
To fix this it was easier to start using BoardDescriptor instead of all
parameters. So I went for it.
Then it seemed I needed access to txt file (internally) which I had
blocked as txt file is not part of the api.
So I extended BoardDescriptor in internalBoardDescriptor made all
constructors of BoardDescriptor Protected and made 2 static create
methods that actually make a InternalBoardDescriptor but return a
BoardDescriptor.
This makes I can safely cast boardDescriptor  to InternalBoardDescriptor
all the time.
Now Teensy Seems to fail