#1167 tested in windows
This fix creates and works with a sloeber variant of boards.txt and
platfrom.txt.
This way the original files are not touched.
This works with previously installed released versions.
Nightly gives issues with some boards due to double workarounds.
Removing and installing again does the trick
The first line in the sloeber variants is used to identify the
"workaround version" to enable updates on the fly

NOTE: this change also copies the package json file to the platform but is not yet using it. Still need to think how to get this to work safely