#1283 first successfull test
I had to do lots of changes to get this to work but now there is a
"convert to sloeber project" which works with a new general project
(when adding a sloeber.cfg and .ino before conversion no additional
actions are needed)
I guess there are plenty of problems but I wanted to check in this code
One of the reasons this was complicated is because config info was
stored based on the ID and not on the name. I neded to switch because tI
do not want the id in the sloeber.cfg and I can not create ID's so I
needed to switch to using names
That also means that there was work on making config name renames work.