#1371 handle tool.x.y expansion based on platform.txt content
Before 4.4 sloeber handled the expansion based on all environment
variables.
In 4.4 this was no longer possible as the variables are not known at the
time when the workaround is applied.
This caused issues
This solution expands based on the platform.txt content itself.
As this is not covering all cases some keywords are always expanded.
Also: some keywords are never expanded