Refactorted and simplified the upload code
Basically I the time I wrote this code I knew even less what I was doing
as I do today.
As there was also some old code that was no longer used I could simplify
the code.
I still think there is a job in a job so it probably still is overly
complicated, but at least it is simpler now