#661 gui and backbone part
We still need to install awk and point to the correct awk location.
We also still need to workaround the ?: problem as this "ignore" method
causes issues in awk
I also only tested avr. The current solution assumes the interesting
data is in the second column. This may not be the case for other
"architectures". I have no clue yet how I will handle this.