Make creation now failes on options #1126
To generate the FLAGS environment variable one needs to know the
"selected options" As I removed the selected option storage from the
model this now faisl.
The good news is; the makefile generation is started; de source file is
found; the tool is found (I assume all makerules are found) and now we
are at the actual makefile generation where the command line generation
fails.