compound project creation #1126
A compound project is a project that can create multiple static libs,
multiple dynamic libs and multiple objects to link them together in one
executable. (What Arduino IDE does -no dynamic libs-; Sloeber currently
only supports 1 static lib)
This works now except for the order of inputs in the link command where
the libs need to be after the objects. Which is one of the reasons why I
started this whole adventure :-)