add cnature projects to the test #1126
Untill now all tests were done with projects that had the  CCnature
nature.
Now tests are added to compile pure C projects.
This means that the nature is now also a parameter for autobuild project
creation code.
I also fixed issues related to the extra tests.
llvm and mvc are the only projects not compiling (42 failures) out of 67
tests