Some refactor rename and make objs build work
objs build did not work for compound project types as the OBJS var was
not given as the var to assign outputs to.
Also I found mingw make 32 had some issues with build when using spaces
in files. Fixed that by replacing spaces with _ for output files and
dependency files
See autobuildCommaon makeNameMakeSafe