Fixed a issue that only happened in maven regression test
Maven failed for the test issue1047_Board_Names_Can_Be_used_as_Strings
The issue related around "-DBOARDID=\"ESP...\""
It worked fine when the test was run from eclipse but failed when
running headless in maven.
Studying the code I found the command is parsed into an array of command
parts (that is the command and the parameters)
I tried lots of things but it only got fixed when the parcing of
parameters no longer dumped the \n
Not sure why it now works both in maven and eclipse.
Let's hope it also work on github now.