Excluded resources from build should not provide include directive #1672
When a library is added to a project the link to the source is added to
the include directive for the build command.
When the library is excluded the include directive should not be on the
build command.
It was ... so I fixed it and then the rename regression test failed.