When converting to sloeber project cope with "wrong" locations
The SLOEBER_HOME folder may have moved. As Sloeber V4 does not do a
great job at identifying SLOEBER_HOMEit is likely (and it was durting
tests) that the boards.txt points to a valid boards.txt (valid in the
sense of: valid path on the original computer and valid if it had been
stored with ${SLOEBER_HOME}
Therefore I added code searching for the ArduinoPlugin folder and go
from there.