Do not recalculate languiage settings when none are available
For optimisation Sloeber caches language settings.
However when the command failed (a list of 0 settings) Sloeber
recalculated these settings.
This was done to allow to refresh the cache however it does cause a
performance hit.
Therefore I opted to accept no results as a valid result untill a better
method is found.