because of #1685 issue1126LibArchiver selects an failing example
Basically #1685 also applies for the library HID-Project which is used
in the test issue1126LibArchiver.
issue1126LibArchiver just selects a "random" example witch due to the
change in #1685 always is a example that does not compile out of the
box.
This change selects the first found example instead of the last found
example.