add tests for #1696 (build should fail)
This commits adds a test for #1696 (actually expanding a existing
tests).
Because the issue dooes not show on linux the test succeeds on my
windows system.
However the build is on a linux system so this commit is expected to
fail the build.
I'm still pushing it to get the test on my Linux system to fix it.
I know I could create a branch and push to the branch ... but as I have
been  the only contributer to this repository the last 12 months nobody
should suffer ;-)